+++ src/main/java/net/minecraft/world/entity/animal/SnowGolem.java
@@ -134,7 +134,7 @@ public class SnowGolem extends AbstractGolem implements Shearable, RangedAttackM
 	@Override
 	protected InteractionResult mobInteract(Player player, InteractionHand interactionHand) {
 		ItemStack itemStack = player.getItemInHand(interactionHand);
-		if (itemStack.is(Items.SHEARS) && this.readyForShearing()) {
+		if (itemStack.is(Items.SHEARS) && !itemStack.isBroken() && this.readyForShearing()) {
 			if (this.level() instanceof ServerLevel serverLevel) {
 				this.shear(serverLevel, SoundSource.PLAYERS, itemStack);
 				this.gameEvent(GameEvent.SHEAR, player);