+++ docs/changes/changes.md
@@ -35,3 +35,4 @@ This Document is a list of all changes in MattMC over base Minecraft Java 1.21.1
 - Swapped Sneak and Sprint keys. Sneak is now LCTRL and Sprint is LSHIFT.
 - Made the GUI Dark Mode
 - Updated whea and potato crop texture to be the same as the Stay True Resource pack. also bushy leaves. Orange Birch leaves.
+- Tools no longer disapear when breaking.